re-implement stories using svelte csf
